+++ Wasteland_Warrior/Wasteland_Warrior/UserInput.cpp
@@ -9,8 +9,8 @@
 std::queue<std::string> UserInput::inputBuffer;
 
 //Mouse
-double UserInput::MouseXpos;
-double UserInput::MouseYpos;
+float UserInput::MouseXpos;
+float UserInput::MouseYpos;
 bool UserInput::MouseLeft;
 bool UserInput::MouseRight;
 
@@ -100,7 +100,7 @@ void UserInput::Update(Gamestate* gameState)
 	//Mouse Input
 	if (UserInput::MouseLeft) {
 		if (UserInput::MouseXpos != oldMouseXpos) {
-			float angle = gameState->cameraAngle + (oldMouseXpos - UserInput::MouseXpos) * 0.01;
+			float angle = gameState->cameraAngle + (oldMouseXpos - UserInput::MouseXpos) * 0.01f;
 			if (angle < 1.5 && angle > -1.5) {
 				gameState->cameraAngle = angle;
 			}
@@ -109,10 +109,10 @@ void UserInput::Update(Gamestate* gameState)
 	}
 	else {
 		if (gameState->cameraAngle > 0.05){
-			gameState->cameraAngle = gameState->cameraAngle - 0.02;
+			gameState->cameraAngle = gameState->cameraAngle - 0.02f;
 		}
 		else if (gameState->cameraAngle < -0.05) {
-			gameState->cameraAngle = gameState->cameraAngle + 0.02;
+			gameState->cameraAngle = gameState->cameraAngle + 0.02f;
 		}
 		else {
 				gameState->cameraAngle = 0;
@@ -241,8 +241,8 @@ void UserInput::key(GLFWwindow* window, int key, int scancode, int action, int m
 
 void UserInput::cursor(GLFWwindow* window, double xpos, double ypos)
 {
-	UserInput::MouseXpos = xpos;
-	UserInput::MouseYpos = ypos;
+	UserInput::MouseXpos = (float)xpos;
+	UserInput::MouseYpos = (float)ypos;
 }
 
 void UserInput::mouseButton(GLFWwindow* window, int button, int action, int mods)
